requirements.txt
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
distlib==0.3.6
filelock==3.8.0
google==3.0.0
google-api-core==2.11.0
google-auth==2.15.0
google-cloud-pubsub==2.13.11
googleapis-common-protos==1.57.1
grpc-google-iam-v1==0.12.4
grpcio==1.51.1
grpcio-status==1.51.1
idna==3.4
platformdirs==2.5.4
proto-plus==1.22.1
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==0.21.0
requests==2.28.1
rsa==4.9
six==1.16.0
soupsieve==2.3.2.post1
urllib3==1.26.12
virtualenv==20.16.7
